move the test where it belongs
